+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Enhancements
 
 Bug Fixes
 - Corrected indeterminate progress dialog repainting
+- Corrected right-aligned columns (e.g. "Last Change") being too narrow to show their full contents
 - Fixed "Automatically elevate on startup" so it works on its own without also disabling the elevation prompt
 
 Miscellaneous